+++ docs/src/global/app.ts
@@ -26,3 +26,4 @@ import '@deckdeckgo/social';
 import '@deckdeckgo/youtube';
 import '@deckdeckgo/drag-resize-rotate';
 import '@deckdeckgo/demo';
+import '@deckdeckgo/math';